Added where to find New-ServicePrincipal
Added small information on how to run New-ServicePrincipal cmdlet, since I've lost some time looking on which module I should install and the only reason it was not available is that I had not enough permission in the tenant I'm connecting to. Instead of having a meaningful error I simply got

```text
new-serviceprincipal : The term 'new-serviceprincipal' is not recognized as the name of a cmdlet, function, script
file, or operable program. Check the spelling of the name, or if a path was included, verify that the path is correct
and try again.
```

leading me to incorrectly believe that I needed to install some other module to find that cmdlet.